bunch of fixes
