+++ backend/tests/test_models/test_system.py
@@ -1,11 +1,13 @@
 import pytest
 
 
-def test_system_attrs():
+
+@pytest.fixture
+def system_cls():
     from models.system import System
-    attrs = ("grace_period", "alarm_count", "ks_enabled", "create_date", "name")
-    assert all(hasattr(System, attr) for attr in attrs) is True
+    return System
 
 
-def test_system_name(random_system):
-    assert random_system.name is not None
+def test_system_attrs(system_cls):
+    attrs = ("grace_period", "alarm_count", "ks_enabled","create_date")
+    assert all(hasattr(system_cls, attr) for attr in attrs) is True
